merge_cachi2_sboms: rename to merge_sboms
Handles syft SBOMs as well now.
