+++ sbom-utility-scripts/Dockerfile
@@ -3,8 +3,10 @@ FROM registry.access.redhat.com/ubi9/python-312:9.5@sha256:88ea2d10c741f16968110
 WORKDIR /scripts
 
 COPY scripts/merge_syft_sboms.py /scripts
-COPY scripts/merge-cachi2-sboms-script/merge_cachi2_sboms.py /scripts
-COPY scripts/merge-cachi2-sboms-script/requirements.txt /scripts/merge-cachi2-sboms-script-requirements.txt
+COPY scripts/merge-sboms-script/merge_sboms.py /scripts
+# for backwards compatibility, also ship the script at the original location
+COPY scripts/merge-sboms-script/merge_sboms.py /scripts/merge_cachi2_sboms.py
+COPY scripts/merge-sboms-script/requirements.txt /scripts/merge-sboms-script-requirements.txt
 COPY scripts/base-images-sbom-script/app/base_images_sbom_script.py /scripts
 COPY scripts/base-images-sbom-script/app/requirements.txt /scripts/base-images-sbom-script-requirements.txt
 COPY scripts/index-image-sbom-script/requirements.txt /scripts/index-image-sbom-script-requirements.txt
@@ -14,7 +16,7 @@ COPY scripts/add-image-reference-script/add_image_reference.py /scripts
 COPY scripts/add-image-reference-script/requirements.txt /scripts/add-image-reference-requirements.txt
 
 RUN pip3 install --no-cache-dir \
-    -r merge-cachi2-sboms-script-requirements.txt \
+    -r merge-sboms-script-requirements.txt \
     -r base-images-sbom-script-requirements.txt \
     -r index-image-sbom-script-requirements.txt \
     -r add-image-reference-requirements.txt
