+++ sbom-utility-scripts/scripts/merge-sboms-script/test_data/cachi2.bom.json
