+++ sbom-utility-scripts/scripts/merge-sboms-script/test_data/syft-sboms/gomod-pandemonium.bom.json
