+++ sbom-utility-scripts/scripts/merge-sboms-script/test_data/syft-sboms/npm-cachi2-smoketest.bom.json
