+++ sbom-utility-scripts/scripts/merge-sboms-script/test_data/syft-sboms/ubi-micro.bom.json
