+++ sbom-utility-scripts/scripts/merge-sboms-script/test_data/syft.merged-by-syft.bom.json
