+++ sbom-utility-scripts/scripts/merge-sboms-script/tox.ini
@@ -6,12 +6,12 @@ basepython = 3.12
 deps =
     -r requirements.txt
     -r requirements-test.txt
-commands = pytest test_merge_cachi2_sboms.py
+commands = pytest test_merge_sboms.py
 
 [testenv:flake8]
 basepython = 3.12
 deps = flake8
-commands = flake8 --max-line-length 120 merge_cachi2_sboms.py test_merge_cachi2_sboms.py
+commands = flake8 --max-line-length 120 merge_sboms.py test_merge_sboms.py
 
 [testenv:black]
 deps = black
